refactor: move project configuration get/setters into a new module
Moves the project configuration frontend code into its own module to
avoid cyclic deps between Project and Obj and keep the Obj module
clean.

I've also removed some noisy comments about "internal only"
project configuration fields.
